requirements.txt
asgiref==3.4.1
attrs==23.1.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.6
django-cors-headers==3.8.0
django-filter==21.1
django-phonenumber-field==6.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
drf-yasg==1.20.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.9.0
idna==3.3
inflection==0.5.1
iniconfig==2.0.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.17.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.41
Pillow==9.0.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.8